Execute source-review Phase 3 (frontend): shared fetch helper, modal extraction, export URL dedup
Frontend-only continuation of docs/SOURCE_REVIEW_COMPLEXITY_SPEED.md section 6
Phase 3. The Kotlin parts (controller error-handling migration, CLI Picocli
consolidation, ConfigBundle/RequirementController de-duplication) need a
working Gradle build to verify safely and are left for a session with that.

- New utils/apiJson.ts: getJson/postJson/putJson/deleteJson wrap the
  authenticated* fetch helpers and return parsed data or throw ApiError with
  the backend's message, replacing the copy-pasted try/ok/catch/finally block
  that existed at ~65 call sites. Sits on top of utils/auth.ts; no token
  handling added. extractErrorMessage moved here from WorkgroupManagement so
  every caller shares one copy of the three-shape error-body parser.

- WorkgroupManagement.tsx (1204 lines) split into a ~300-line table/launcher
  parent plus three owned dialogs: WorkgroupFormModal, WorkgroupAssignUsersModal,
  WorkgroupAssignAssetsModal. Shared shapes moved to workgroupTypes.ts and the
  pure assign-logic (effective-count math, wildcard asset filter) to
  workgroupAssignLogic.ts, both now unit-tested under the Node test tier.

- Export.tsx and ImportExport.tsx: their four duplicated docx/xlsx endpoint
  builders replaced with one buildAuthenticatedRequirementExportUrl helper
  (extends the existing requirementDownloadUrl.ts). This also fixes a real
  drift: ImportExport's use-case xlsx export had silently dropped the
  releaseId parameter. Error-body parsing in both files now goes through the
  shared extractErrorMessage instead of four inline copies.

Verified: npm ci && npm run build clean, 274/0 unit tests (17 new),
owasp-check 0 findings, humanizer-scan 0 BLOCK / 7 REVIEW (all pre-existing
FUNC-MAX on Export/ImportExport/WorkgroupManagement — each shrank this pass
but a full component split is deferred; noted, not silently skipped).
